Fix notification for removing responses

See merge request c1718515/swiperightbrains!57

+++ src/main/java/com/nsa/cubric/application/controllers/API/AdminAPI.java
@@ -36,11 +36,11 @@ public class AdminAPI {
     }
 
     @PostMapping(value = "/removeUserResponses")
-    public Long removeUserResponses(@RequestParam(value = "userToRemoveResponses") String userEmail){
+    public Integer removeUserResponses(@RequestParam(value = "userToRemoveResponses") String userEmail){
         try{
             return adminServices.removeUserResponses(accountService.getAccountByEmail(userEmail).getId());
         } catch (NullPointerException e){
-            return 0L;
+            return 0;
         }
     }
